+++ src/theme/container.ts
@@ -5,5 +5,5 @@
  */
 
 export default {
-  base: 'w-full max-w-[1280px] mx-auto px-[22px]' // max-w-7xl w-full
+  base: 'w-full max-w-(--b24ui-container-width) mx-auto px-4 sm:px-6 lg:px-8'
 }